# server/swagger/swagger.yaml
definitions:
...
  api.Meta:
    properties:
      error:
        type: string
      msg:
        description: Status int64  `json:"status"`
        type: string
    type: object
...
  api.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.UpdateDomainAssHostReq:
    properties:
      did:
        type: integer
      hids:
        items:
          type: integer
        type: array
    required:
    - did
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/host/Host:
    get:
      description: 全部查询不传Ip
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 查询IP则输入IP，v4或v6都可以
        in: query
        name: ip
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 查询服务器
      tags:
      - 服务器相关
...
  /api/v1/host/assDomain:
    put:
      description: 服务器ID[多选]
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关联传入参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UpdateDomainAssHostReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 域名关联服务器
      tags:
      - 服务器相关
...
  /api/v1/host/domainAssHost:
    get:
      description: 返回服务器切片
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 查询域名对应服务器
      tags:
      - 服务器相关
  /api/v1/host/update:
    post:
      description: 返回新增/修改的指定服务器
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 1 按量收费, 2 包月收费, 3 包年收费 ...后续有需要再加
        in: formData
        name: billing
        required: true
        type: integer
      - in: formData
        name: cloud
        required: true
        type: string
      - description: 四核输入4
        in: formData
        name: cores
        required: true
        type: integer
      - description: 下次续费金额, 人民币为单位
        in: formData
        name: cost
        type: number
      - description: 磁盘单位为G
        in: formData
        name: data_disk
        required: true
        type: integer
      - in: formData
        name: id
        type: integer
      - in: formData
        name: iops
        required: true
        type: integer
      - in: formData
        name: ipv4
        required: true
        type: string
      - in: formData
        name: ipv6
        type: string
      - in: formData
        name: mbps
        required: true
        type: integer
      - description: 内存单位为G
        in: formData
        name: mem
        required: true
        type: integer
      - collectionFormat: csv
        in: formData
        items:
          type: integer
        name: password
        type: array
      - in: formData
        name: port
        required: true
        type: string
      - in: formData
        name: system
        required: true
        type: string
      - description: 磁盘单位为G
        in: formData
        name: system_disk
        required: true
        type: integer
      - description: 1 单服机器, 2 中央服机器, 3 CDN机器, 4 业务服机器  ...后续有需要再加
        in: formData
        name: type
        required: true
        type: integer
      - in: formData
        name: user
        required: true
        type: string
      - description: 所在地
        in: formData
        name: zone
        required: true
        type: string
      - description: 时区，如东八区填8
        in: formData
        name: zone_time
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 新增/修改服务器
      tags:
      - 服务器相关
  /api/v1/host/updateDomain:
    post:
      description: 返回新增或修改的指定域名
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: domain
        required: true
        type: string
      - in: formData
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 新增/修改域名
      tags:
      - 服务器相关
...
  /api/v1/project/update:
    post:
      description: 返回新增/修改的指定项目
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: group_id
        required: true
        type: integer
      - description: 修改才需要传，没有传算新增
        in: formData
        name: id
        type: integer
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: status
        required: true
        type: integer
      - in: formData
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 新增/修改项目
      tags:
      - 项目相关
...
swagger: "2.0"